+++ com.sparta-express.order-shipment/src/main/java/com/sparta_express/order_shipment/presentation/controller/ShipmentRouteController.java
@@ -3,15 +3,13 @@ package com.sparta_express.order_shipment.presentation.controller;
 import com.querydsl.core.types.Predicate;
 import com.sparta_express.order_shipment.application.dto.*;
 import com.sparta_express.order_shipment.application.service.ShipmentRouteService;
-import com.sparta_express.order_shipment.common.PageableUtil;
+import com.sparta_express.order_shipment.common.util.PageableUtil;
 import com.sparta_express.order_shipment.domain.entity.ShipmentRoute;
 import com.sparta_express.order_shipment.presentation.dto.ShipmentRouteUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.querydsl.binding.QuerydslPredicate;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
